+++ pakyow-data/spec/failures/unknown_subscriber_adapter_spec.rb
@@ -7,12 +7,16 @@ RSpec.describe "using an unknown subscribers adapter" do
     end
   end
 
-  let :allow_application_rescues do
-    true
-  end
+  let(:autorun) {
+    false
+  }
 
-  it "boots the app in rescue mode" do
-    expect(Pakyow.apps.first.rescued?).to be(true)
-    expect(call("/")[2]).to include("Failed to load subscriber adapter named `foo'")
+  it "raises an error" do
+    expect {
+      setup_and_run
+    }.to raise_error(Pakyow::ApplicationError) do |error|
+      expect(error.cause).to be_instance_of(Pakyow::Data::UnknownSubscriberAdapter)
+      expect(error.message).to eq("failed to load subscriber adapter named `foo'")
+    end
   end
 end